+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Emogrifier Change Log
 
 All notable changes to this project will be documented in this file.
-This project adheres to [Semantic Versioning](http://semver.org/).
+This project adheres to [Semantic Versioning](https://semver.org/).
 
 
-## x.y.z (unreleased)
+## 2.0.0
 
 ### Added
 - Support for CSS :not() selector